refactor(desktop): split team profile bridge into its own module
Move settings-view and reset-field JSON-RPC handling out of
desktop_backend.rs so the v0.36.26 desktop split gate stays within
budget. Register desktop_team_profile.rs in the gate module list.
