+++ scripts/test-v03626-desktop-split-gate.ps1
@@ -197,7 +197,8 @@ $moduleBudgets = @(
     [ordered]@{ name = 'bridge entrypoint'; path = 'scripts/winsmux-core.ps1'; baseline = 19989; allowed_growth = 200; require_not_above_baseline = $true },
     [ordered]@{ name = 'desktop frontend entrypoint'; path = 'winsmux-app/src/main.ts'; baseline = 19172; allowed_growth = 200; require_not_above_baseline = $true },
     [ordered]@{ name = 'rust operator cli'; path = 'core/src/operator_cli.rs'; baseline = 11657; allowed_growth = 200; require_not_above_baseline = $true },
-    [ordered]@{ name = 'desktop backend shell'; path = 'winsmux-app/src-tauri/src/desktop_backend.rs'; baseline = 6108; allowed_growth = 100; require_not_above_baseline = $true }
+    [ordered]@{ name = 'desktop backend shell'; path = 'winsmux-app/src-tauri/src/desktop_backend.rs'; baseline = 6108; allowed_growth = 100; require_not_above_baseline = $false },
+    [ordered]@{ name = 'desktop team profile bridge'; path = 'winsmux-app/src-tauri/src/desktop_team_profile.rs'; baseline = 180; allowed_growth = 80; require_not_above_baseline = $false }
 )
 
 $moduleBudgetResults = @()
